backend/requirements.txt
Django==4.2.25
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
whitenoise==6.6.0

# Database
psycopg2-binary==2.9.9

# Cloud pricing APIs
boto3==1.34.162
requests==2.31.0
google-cloud-billing==1.12.3
google-auth==2.23.4
google-oauth2-tool==0.0.3

# Additional utilities
celery==5.3.4
redis==5.0.1